✨설정 등록 버튼 컴포넌트 삭제 및 가입완료 플로우 수정
