+++ pages/oauth/_components/SignUpModal.tsx
@@ -1,9 +1,9 @@
+import CelebrateRegister from 'components/initialRegister/CelebrateRegister';
 import { useRouter } from 'next/router';
 import { VFC } from 'react';
 
 import Modal from '../../../components/common/Modal';
 import ModalTemplate from '../../../components/common/Modal/ModalTemplate';
-import SetUserProfile from '../../../components/initialRegister/SetUserProfile';
 
 interface Props {
   isShowing: boolean;
@@ -20,7 +20,7 @@ const SignUpModal: VFC<Props> = ({ isShowing }) => {
     <>
       <Modal isShowing={isShowing} hide={setModalVisible}>
         <ModalTemplate hide={setModalVisible} showDelete={false}>
-          <SetUserProfile />
+          <CelebrateRegister />
         </ModalTemplate>
       </Modal>
     </>
